Improve TipoPrestador DAO and Business classes

+++ agenda-digital-dao/src/main/java/com/everis/academia/java/agenda/digital/dao/inter/ITipoServicoDAO.java
@@ -33,7 +33,7 @@ public interface ITipoServicoDAO {
 	 * 
 	 * @param codigo
 	 */
-	void delete(Integer codigo);
+	void delete(Short codigo);
 	
 	/**
 	 * Verifica se já existe um Tipo de Serviço
